Work in progress to implement a generalized Fixed-Point model

+++ sources/libdetectors/CDetector/CScoreProcessor/CNorm/CNorm.cpp
@@ -5,5 +5,5 @@ class phantom : public QCSP::StandaloneDetector::CNorm<64, float> {};
 } // namespace
 
 namespace {
-class phantom_int : public QCSP::StandaloneDetector::CNorm<64, int> {};
+class phantom_int : public QCSP::StandaloneDetector::CNorm<64, int16_t> {};
 } // namespace
